Move updateArray back to observable-array module
